# openapi/newton-backend-parity.yaml
openapi: 3.1.0
info:
  title: Newton Backend Parity API
  description: Contract for Rust `newton serve` parity endpoints sourced from Rust DTOs and handlers.
  contact:
    name: Newton Loop Framework
  license:
    name: ''
  version: 0.5.101
servers:
- url: /api/v1
paths:
...
  /workflows:
    get:
      tags:
      - workflows
      operationId: list_workflows
      parameters:
      - name: status
        in: query
        description: Optional workflow status filter
        required: false
        schema:
          $ref: '#/components/schemas/WorkflowStatus'
      - name: limit
        in: query
        description: Maximum number of workflow instances
        required: false
        schema:
          type: integer
          minimum: 0
      - name: offset
        in: query
        description: Number of workflow instances to skip
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Workflow instance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowInstance'
    post:
      tags:
      - workflows
      operationId: create_workflow
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowInstance'
        required: true
      responses:
        '201':
          description: Created workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstance'
        '409':
          description: Workflow instance already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /workflows/{id}:
    get:
      tags:
      - workflows
      operationId: get_workflow
      parameters:
      - name: id
        in: path
        description: Workflow instance id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstance'
        '404':
          description: Workflow instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      tags:
      - workflows
      operationId: update_workflow
      parameters:
      - name: id
        in: path
        description: Workflow instance id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateBody'
        required: true
      responses:
        '200':
          description: Updated workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstance'
        '404':
          description: Workflow instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /workflows/{id}/nodes/{node_id}:
    patch:
      tags:
      - workflows
      operationId: update_node
      parameters:
      - name: id
        in: path
        description: Workflow instance id
        required: true
        schema:
          type: string
      - name: node_id
        in: path
        description: Workflow node id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeUpdate'
        required: true
      responses:
        '200':
          description: Updated workflow instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstance'
        '404':
          description: Workflow instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - category
      - message
      properties:
        category:
          type: string
        code:
          type: string
        details: {}
        message:
          type: string
...
    NodeState:
      type: object
      required:
      - node_id
      - status
      properties:
        ended_at:
          type:
          - string
          - 'null'
          format: date-time
        node_id:
          type: string
        operator_type:
          type:
          - string
          - 'null'
        started_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          $ref: '#/components/schemas/NodeStatus'
    NodeStatus:
      type: string
      enum:
      - pending
      - running
      - succeeded
      - failed
      - timeout
      - cancelled
    NodeUpdate:
      type: object
      required:
      - status
      properties:
        ended_at:
          type:
          - string
          - 'null'
          format: date-time
        operator_type:
          type:
          - string
          - 'null'
        started_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          $ref: '#/components/schemas/NodeStatus'
...
    WorkflowInstance:
      type: object
      required:
      - instance_id
      - workflow_id
      - status
      - nodes
      - started_at
      properties:
        definition: {}
        ended_at:
          type:
          - string
          - 'null'
          format: date-time
        instance_id:
          type: string
        linked_plan_id:
          type:
          - string
          - 'null'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeState'
        started_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        workflow_id:
          type: string
    WorkflowStatus:
      type: string
      enum:
      - running
      - succeeded
      - failed
      - paused
      - cancelled
    WorkflowUpdateBody:
      type: object
      description: |-
        Flexible update body: supports both legacy WorkflowDefinition format
        and new status/ended_at update format.
      properties:
        definition: {}
        ended_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/WorkflowStatus'
        workflow_id:
          type:
          - string
          - 'null'
...